### 2018-05-15 #734 @whedon
[PRE REVIEW]: SpeechPy - Speech Processing and Recognition
**Submitting author:** @astorfi (<a href="http://orcid.org/0000-0003-2282-4361">Amirsina Torfi</a>)
**Repository:** <a href="https://github.com/astorfi/speechpy" target ="_blank">https://github.com/astorfi/speechpy</a>
**Version:** 2.3
