### 2018-05-15 #734 @arfon
👋@astorfi - please take a look at the submission guidelines and format your paper _exactly_ as outlined here: http://joss.theoj.org/about#paper_structure
