+++ paper/paper.bib
@@ -1,0 +1,37 @@
+@article{torfi2017text,
+  title={Text-independent speaker verification using 3d convolutional neural networks},
+  author={Torfi, Amirsina and Nasrabadi, Nasser M and Dawson, Jeremy},
+  journal={arXiv preprint arXiv:1705.09422},
+  year={2017}
+}
+
+@article{torfi20173d,
+  title={3D Convolutional Neural Networks for Cross Audio-Visual Matching Recognition},
+  author={Torfi, Amirsina and Iranmanesh, Seyed Mehdi and Nasrabadi, Nasser and Dawson, Jeremy},
+  journal={IEEE Access},
+  volume={5},
+  pages={22081--22091},
+  year={2017},
+  publisher={IEEE}
+}
+
+@article{torfi2018speechpy,
+  title={Speechpy-a library for speech processing and recognition},
+  author={Torfi, Amirsina},
+  journal={arXiv preprint arXiv:1803.01094},
+  year={2018}
+}
+
+@article{mobiny2018text,
+  title={Text-Independent Speaker Verification Using Long Short-Term Memory Networks},
+  author={Mobiny, Aryan},
+  journal={arXiv preprint arXiv:1805.00604},
+  year={2018}
+}
+
+@article{salehghaffari2018speaker,
+  title={Speaker Verification using Convolutional Neural Networks},
+  author={Salehghaffari, Hossein},
+  journal={arXiv preprint arXiv:1803.05427},
+  year={2018}
+}
