### 2018-05-16 #734 @astorfi
@arfon @whedon I made some modifications. Please check to see if the problem is resolved. However, the [paper.md](https://github.com/astorfi/speechpy/blob/master/paper/paper.md) file is same as before. I do not know what is the issue and what should be changed because to my view it is exactly as outlined.
