### 2018-05-17 #734 @arfon
Hi @astorfi.

Please compare this (yours):

```
---
title: 'SpeechPy - Speech Processing and Recognition'
tags:
  - Python
authors:
  - name: Amirsina Torfi
    orcid: 0000-0003-2282-4361
    affiliation: "Virginia Tech"
---
```
with this (the documented example):

```
---
title: 'Gala: A Python package for galactic dynamics'
tags:
  - Python
  - astronomy
  - dynamics
  - galactic dynamics
  - milky way
authors:
  - name: Adrian M. Price-Whelan
    orcid: 0000-0003-0872-7098
    affiliation: "1, 2" # (Multiple affiliations must be quoted)
  - name: Author 2
    orcid: 0000-0000-0000-0000
    affiliation: 2
affiliations:
 - name: Lyman Spitzer, Jr. Fellow, Princeton University
   index: 1
 - name: Institution 2
   index: 2
date: 13 August 2017
bibliography: paper.bib
---
```
