### 2018-05-17 #734 @arfon
> @arfon Thank you so much. It's been modified. Please check for any possible issues.

Thanks @astorfi. Could you take a look at this list of reviewers and suggest a few people who could potentially review this submission? https://bit.ly/joss-reviewers
