### 2018-05-18 #734 @arfon
@jjatria @josephhardinee would you be willing to review this submission for JOSS?
