### 2018-05-18 #734 @jjatria
@arfon It seems like @josephhardinee might be more proficient than I am on Python, but I'll be happy to take a look. Does the review process normally involve a single reviewer?

I know another researcher who might be the ideal person to review this, but I'm not sure if they are available / interested...
