### 2018-05-20 #734 @arfon
> I know another researcher who might be the ideal person to review this, but I'm not sure if they are available / interested...

Sure thing. If you're comfortable, please mention them here (with their GitHub handle).

> Does the review process normally involve a single reviewer?

Ideally we have more than one reviewer, especially when the topic area is outside of the editor's specialism.
