### 2018-06-28 #749 @ThomasA
I checked the article proof. Comparing to the GitHub-formatted Markdown [here](https://github.com/astorfi/speechpy/blob/master/paper/paper.md), the PDF seems to be typeset wrong: the bullet list specified with "+"-es does not get typeset properly.
