### 2018-06-28 #749 @ThomasA
If you don't mind, I am just going to comment bit by bit here as I go through the review...

- Version: Links to the repository do not point to any specific version, but there is a release tagged 2.3 in the repository. I suppose this is OK?
