### 2018-06-28 #749 @ThomasA
- Authorship: In addition to @astorfi, @MatthewScholefield has made a small contribution to the project and @cclauss a very tiny one. Given that @astorfi has contributed the vast majority of the package, I guess it is OK that he is the sole author of this publication?
