### 2018-06-28 #749 @ThomasA
- Installation: This works as documented in the repository's README, but: I explicitly checked out the '2.3' release for installation whereas the install script output refers to version 2.2:

    $ python setup.py develop
    running develop
    running egg_info
    creating speechpy.egg-info
    writing speechpy.egg-info/PKG-INFO
    writing dependency_links to speechpy.egg-info/dependency_links.txt
    writing requirements to speechpy.egg-info/requires.txt
    writing top-level names to speechpy.egg-info/top_level.txt
    writing manifest file 'speechpy.egg-info/SOURCES.txt'
    reading manifest file 'speechpy.egg-info/SOURCES.txt'
    reading manifest template 'MANIFEST.in'
    writing manifest file 'speechpy.egg-info/SOURCES.txt'
    running build_ext
    Creating /home/tha/.conda/envs/joss-review-tmp/lib/python3.6/site-packages/speechpy.egg-link (link to .)
    Adding speechpy 2.2 to easy-install.pth file

    Installed /home/tha/other-repo/speechpy
    Processing dependencies for speechpy==2.2
    Searching for numpy==1.14.3
    Best match: numpy 1.14.3
    Adding numpy 1.14.3 to easy-install.pth file

    Using /home/tha/.conda/envs/joss-review-tmp/lib/python3.6/site-packages
    Searching for scipy==1.1.0
    Best match: scipy 1.1.0
    Adding scipy 1.1.0 to easy-install.pth file

    Using /home/tha/.conda/envs/joss-review-tmp/lib/python3.6/site-packages
    Finished processing dependencies for speechpy==2.2

I suspect this is just some configuration text string that was not properly updated for the 2.3 release, but I think @astorfi should check and possibly correct this so we are sure that it indeed installs the 2.3 release.
