### 2018-06-28 #749 @ThomasA
- Example usage: Glancing through the documentation, I did not find any usage examples, but the test scripts can be considered not-too-well-documented examples. A few usage examples that were easier to find and documented with a bit of explanatory text would be nice to have.
