### 2018-06-28 #749 @ThomasA
- Functionality documentation: The package's functions are documented with docstrings and nicely formatted in the package HTML documentation. The modules of the package, however, do not have any docstrings. This could be a slight improvement, also when quickly looking up documentation from IPython, for example: `speechpy.processing?`

      Type:        module
      String form: <module 'speechpy.processing' from '/home/tha/other-repo/speechpy/speechpy/processing.py'>
      File:        ~/other-repo/speechpy/speechpy/processing.py
      Docstring:   <no docstring>
