### 2018-06-28 #749 @ThomasA
- Automated tests: There are tests available in the 'tests' folder. These are in my opinion more like usage examples and not proper tests in the sense that they do not check anything. They run some example commands and display the output. In my opinion, proper automated tests should check the output and determine whether it is correct, resulting in clearly 'passed' or 'failed' test cases.
