### 2018-06-28 #749 @ThomasA
- Community guidelines: The package documentation has very clear guidelines on [how to contribute](http://speechpy.readthedocs.io/en/latest/epilogue/finalnote.html#contributing) to the project. It is however a bit unclear on the reporting issues and seeking support aspects. The latter two aspects could be improved for example by mentioning the that issues should be reported in the repository's issue tracker or similar.

  In addition, it would be helpful to mention these things in the repository's README as well, where potential users and contributors can more easily find it at a quick glance.
