### 2018-06-28 #749 @ThomasA
- References: The software paper does not list any references and as such does not need DOIs for these.
