### 2018-06-28 #749 @astorfi
@ThomasA Thank you for your review. I will investigate them and will comment here.
