### 2018-06-28 d7a152f
add references
