+++ setup.py
@@ -5,8 +5,8 @@ setup(name='speechpy',
       description='The python package for extracting speech features.',
       author='Amirsina Torfi',
       author_email='amirsina.torfi@gmail.com',
-      url='https://github.com/astorfi/speechpyn',
-      download_url = 'https://github.com/astorfi/speechpy/archive/2.2.zip',
+      url='https://github.com/astorfi/speechpy',
+      download_url = 'https://github.com/astorfi/speechpy/archive/2.3.zip',
       packages=find_packages(exclude=('tests', 'docs')),
       include_package_data=True,
       install_requires=[
