### 2018-06-28 #749 @astorfi
@ThomasA About the docstring for modules you are absolutely correct. I am modifying the repo as you suggested and will investigate the rest afterwards.
