### 2018-06-28 #749 @astorfi
@ThomasA Just added docstrings for modules. Please kindly send me your feedback.
