### 2018-06-28 #749 @astorfi
@ThomasA @arfon About the "+" es, I just check it and you are correct. I changed them to "*" es with some space. However, I do not know how to render PDF to see if it resolved the issue.
