### 2018-06-28 938b41b
add references
