+++ docs/source/epilogue/CONTRIBUTING.rst
@@ -1,0 +1,31 @@
+
+*************
+Contributing
+*************
+
+When contributing to this repository, you are more than welcome to discuss your feedback with any of the owners of this repository. *For typos, please do not create a pull request. Instead, declare them in issues or email the repository owner*. For technical and conceptual questions please feel free to **directly contact the repository owner**. Before asking general questions related to the concepts and techniques provided in this project, **please make sure to read and understand its associated paper**. 
+
+Please note we have a code of conduct, please follow it in all your interactions with the project.
+
+====================
+Pull Request Process
+====================
+
+Please consider the following criterions in order to help us in a better way:
+
+1. The pull request is mainly expected to be a code script suggestion or improvement.
+2. A pull request related to non-code-script sections is expected to make a significant difference in the documentation. Otherwise, it is expected to be announced in the issues section.
+3. Ensure any install or build dependencies are removed before the end of the layer when doing a 
+   build and creating a pull request.
+4. Add comments with details of changes to the interface, this includes new environment 
+   variables, exposed ports, useful file locations and container parameters.
+5. You may merge the Pull Request in once you have the sign-off of at least one other developer, or if you 
+   do not have permission to do that, you may request the owner to merge it for you if you believe all checks are passed.
+
+============
+Final Note
+============
+
+We are looking forward to your kind feedback. Please help us to improve this open source project and make our work better. 
+For contribution, please create a pull request and we will investigate it promptly. Once again, we appreciate 
+your kind feedback and elaborate code inspections.
