### 2018-06-28 af321bf
add references
