### 2018-06-28 afb27c2
add references
