### 2018-06-28 90826c2
add references
