### 2018-06-28 bf8efdf
add references
