+++ README.rst
@@ -20,9 +20,9 @@
 .. _SpeechPy Official Project Documentation: http://speechpy.readthedocs.io
 
 
-============================================
+##################
 Table of Contents
-============================================
+##################
 .. contents::
 :local:
 :depth: 3
