+++ README.rst
@@ -24,8 +24,8 @@
 Table of Contents
 ==========================
 .. contents::
-:local:
-:depth: 2
+  :local:
+  :depth: 3
 
 
 
