### 2018-06-28 #749 @astorfi
@ThomasA As you kindly suggested, I added a contributing section in the README as well as how to declare and follow issues. However, about ``seeking support aspects`` that you mentioned, I need a little bit more clarification if possible.
