+++ docs/source/conf.py
@@ -20,6 +20,7 @@ import os
 import sys
 sys.path.insert(0, os.path.abspath('../../'))
 import speechpy
+import numpy
 import sphinx_rtd_theme
 
 
