+++ docs/source/conf.py
@@ -141,7 +141,7 @@ htmlhelp_basename = 'SpeechPydoc'
 
 # latex_engine = 'pdflatex'
 
-latex_engine = 'xelatex'
+latex_engine = 'lualatex'
 latex_elements = {
 
 'papersize': 'a4paper',
