+++ docs/source/conf.py
@@ -141,28 +141,92 @@ htmlhelp_basename = 'SpeechPydoc'
 
 # latex_engine = 'pdflatex'
 
-latex_engine = 'lualatex'
-latex_elements = {
+# latex_engine = 'lualatex'
+# latex_elements = {
 
-'papersize': 'a4paper',
-    'releasename':" ",
-    'figure_align':'htbp',
-    'pointsize': '12pt',
-    'fontpkg': r'''
-\setmainfont{Times New Roman}
-\setsansfont{Times New Roman}
-\setmonofont{Times New Roman}
-''',
-    'preamble': r'''
-\usepackage[titles]{tocloft}
-\cftsetpnumwidth {1.25cm}\cftsetrmarg{1.5cm}
-\setlength{\cftchapnumwidth}{0.75cm}
-\setlength{\cftsecindent}{\cftchapnumwidth}
-\setlength{\cftsecnumwidth}{1.25cm}
-''',
-    'fncychap': r'\usepackage[Bjornstrup]{fncychap}',
-    'printindex': r'\footnotesize\raggedright\printindex',
+# 'papersize': 'a4paper',
+#     'releasename':" ",
+#     'figure_align':'htbp',
+#     'pointsize': '12pt',
+#     'fontpkg': r'''
+# \setmainfont{Times New Roman}
+# \setsansfont{Times New Roman}
+# \setmonofont{Times New Roman}
+# ''',
+#     'preamble': r'''
+# \usepackage[titles]{tocloft}
+# \cftsetpnumwidth {1.25cm}\cftsetrmarg{1.5cm}
+# \setlength{\cftchapnumwidth}{0.75cm}
+# \setlength{\cftsecindent}{\cftchapnumwidth}
+# \setlength{\cftsecnumwidth}{1.25cm}
+# ''',
+#     'fncychap': r'\usepackage[Bjornstrup]{fncychap}',
+#     'printindex': r'\footnotesize\raggedright\printindex',
+# }
+
+
+
+
+latex_elements = {
+    # The paper size ('letterpaper' or 'a4paper').
+    #
+    'papersize': 'letterpaper',
+
+    # The font size ('10pt', '11pt' or '12pt').
+    #
+    'pointsize': '10pt',
+
+    # Additional stuff for the LaTeX preamble.
+    #
+    'preamble': '',
+
+    # Latex figure (float) alignment
+    #
+    'figure_align': 'htbp',
 }
+
+# Grouping the document tree into LaTeX files. List of tuples
+# (source start file, target name, title,
+#  author, documentclass [howto, manual, or own class]).
+latex_documents = [
+    (master_doc, 'test.tex', u'test Documentation',
+     u'test', 'manual'),
+]
+
+# The name of an image file (relative to this directory) to place at the top of
+# the title page.
+#
+# latex_logo = None
+
+# If true, show page references after internal links.
+#
+# latex_show_pagerefs = False
+
+# If true, show URL addresses after external links.
+#
+# latex_show_urls = False
+
+# Documents to append as an appendix to all manuals.
+#
+# latex_appendices = []
+
+# If false, no module index is generated.
+#
+# latex_domain_indices = True
+
+
+
+
+
+
+
+
+
+
+
+
+
+
 
 latex_logo = '_static/img/speechpy_logo.jpg'
 
@@ -325,7 +389,7 @@ latex_logo = '_static/img/speechpy_logo.jpg'
 # (source start file, target name, title,
 #  author, documentclass [howto, manual, or own class]).
 latex_documents = [
-    (master_doc, 'SpeechPy.tex', 'SpeechPy Documentation',
+    (master_doc, 'speechpy.tex', 'SpeechPy Documentation',
      'Amirsina Torfi', 'manual'),
 ]
 
