### 2018-06-28 #749 @astorfi
@ThomasA  About creating test cases, I am working on it and will get back with the updated information.
