### 2018-06-29 #749 @ThomasA
@arfon more general question; how is it handled when a review leads to some revisions - like here? The tagged version 2.3 of the repository will be out-dated by the revisions discussed here. Should the collected revisions be released as 2.3.1 and the paper here updated to refer to that version?
