### 2018-06-29 #749 @ThomasA
Yay, it works! @astorfi the bullet list comes out right now.
