### 2018-06-29 #749 @ThomasA
I need to be careful not to get too picky here, but I can perhaps recommend you run `pycodestyle` on your modules to get [PEP8](https://www.python.org/dev/peps/pep-0008/) code style recommendations? Over-long lines of text in docstrings are a bit annoying to read in terminals...
