### 2018-06-29 e91fdcd
Travis CI: Run flake8 across the codebase
