### 2018-06-29 #749 @cclauss
https://github.com/astorfi/speechpy/pull/19 runs http://flake8.pycqa.org testing which is a superset of PyCodeStyle (fka PEP8) + PyFlakes + McCabe complexity + syntax checking.
