### 2018-06-29 #749 @ThomasA
@cclauss sounds like a good idea. Fixing for example over-long lines would be nice.
