### 2018-06-29 af3e947
Some whitespace fixes
