+++ tests/test_package.py
@@ -3,16 +3,17 @@ import numpy as np
 import speechpy
 import os
 
-file_name = os.path.join(os.path.dirname(os.path.abspath(__file__)),'Alesis-Sanctuary-QCard-AcoustcBas-C2.wav')
+file_name = os.path.join(os.path.dirname(os.path.abspath(__file__)),
+                         'Alesis-Sanctuary-QCard-AcoustcBas-C2.wav')
 fs, signal = wav.read(file_name)
-signal = signal[:,0]
+signal = signal[:, 0]
 
 # Example of pre-emphasizing.
 signal_preemphasized = speechpy.processing.preemphasis(signal, cof=0.98)
 
 # Example of staching frames
-frames = speechpy.processing.stack_frames(signal, sampling_frequency=fs, frame_length=0.020, frame_stride=0.01, filter=lambda x: np.ones((x,)),
-         zero_padding=True)
+frames = speechpy.processing.stack_frames(signal, sampling_frequency=fs, frame_length=0.020,
+             frame_stride=0.01, filter=lambda x: np.ones((x,)), zero_padding=True)
 
 # Example of extracting power spectrum
 power_spectrum = speechpy.processing.power_spectrum(frames, fft_points=512)
@@ -32,7 +33,3 @@ logenergy = speechpy.feature.lmfe(signal, sampling_frequency=fs, frame_length=0.
              num_filters=40, fft_length=512, low_frequency=0, high_frequency=None)
 logenergy_feature_cube = speechpy.feature.extract_derivative_feature(logenergy)
 print('logenergy features=', logenergy.shape)
-
-
-
-
