### 2018-06-29 #749 @arfon
> @arfon more general question; how is it handled when a review leads to some revisions - like here? The tagged version 2.3 of the repository will be out-dated by the revisions discussed here. Should the collected revisions be released as 2.3.1 and the paper here updated to refer to that version?

Yes, we'll just update the revision at the end of the review (when the archive of the code is made).
