### 2018-06-29 #749 @ThomasA
> Just added docstrings for modules. Please kindly send me your feedback.

I have proposed some minor revisions of your docstrings in https://github.com/astorfi/speechpy/pull/21.
