### 2018-06-29 #749 @ThomasA
> However, about seeking support aspects that you mentioned, I need a little bit more clarification if possible.

I think you have got this covered by "technical and conceptual questions".
