### 2018-06-29 #749 @astorfi
@ThomasA  Thank you for all the suggestions. I will investigate them and will get back with the possible changes.
