+++ speechpy/processing (copy).py
@@ -1,14 +1,14 @@
 # -*- coding: utf-8 -*-
 """Processing module for signal processing operations.
 
-This module demonstrates documentation for the signal processing
-function which are required as internal computations in the package.
+This module demonstrates documentation for the signal processing function which are 
+required as internal computations in the package.
 
 
 Attributes:
 
     preemphasis: Preemphasising on the signal. This is a preprocessing step.
-
+    
     stack_frames: Create stacking frames from the raw signal.
 
     fft_spectrum: Calculation of the Fast Fourier Transform.
@@ -17,14 +17,11 @@ Attributes:
 
     log_power_spectrum: Log Power Spectrum calculation.
 
-    derivative_extraction: Calculation of the derivative
-        of the extracted featurs.
+    derivative_extraction: Calculation of the derivative of the extracted featurs.
 
-    cmvn: Cepstral mean variance normalization.
-        This is a post processing operation.
+    cmvn: Cepstral mean variance normalization. This is a post processing operation.
 
-    cmvnw: Cepstral mean variance normalization over the sliding window.
-        This is a post processing operation.
+    cmvnw: Cepstral mean variance normalization over the sliding window. This is a post processing operation.
 
 """
 
@@ -39,10 +36,7 @@ import math
 
 # 1.4 becomes 1 and 1.6 becomes 2. special case: 1.5 becomes 2.
 def round_half_up(number):
-    return int(
-        decimal.Decimal(number).quantize(
-            decimal.Decimal('1'),
-            rounding=decimal.ROUND_HALF_UP))
+    return int(decimal.Decimal(number).quantize(decimal.Decimal('1'), rounding=decimal.ROUND_HALF_UP))
 
 
 def preemphasis(signal, shift=1, cof=0.98):
@@ -60,16 +54,8 @@ def preemphasis(signal, shift=1, cof=0.98):
     rolled_signal = np.roll(signal, shift)
     return signal - cof * rolled_signal
 
-
-def stack_frames(
-        sig,
-        sampling_frequency,
-        frame_length=0.020,
-        frame_stride=0.020,
-        filter=lambda x: np.ones(
-            (x,
-             )),
-        zero_padding=True):
+def stack_frames(sig, sampling_frequency, frame_length=0.020, frame_stride=0.020, filter=lambda x: np.ones((x,)),
+                 zero_padding=True):
     """Frame a signal into overlapping frames.
 
     Args:
@@ -77,34 +63,27 @@ def stack_frames(
         sampling_frequency (int): The sampling frequency of the signal.
         frame_length (float): The length of the frame in second.
         frame_stride (float): The stride between frames.
-        filter (array): The time-domain filter for applying to each frame.
-            By default it is one so nothing will be changed.
-        zero_padding (bool): If the samples is not a multiple of
-            frame_length(number of frames sample), zero padding will
-            be done for generating last frame.
+        filter (array): The time-domain filter for applying to each frame. By default it is one so nothing will be changed.
+        zero_padding (bool): If the samples is not a multiple of frame_length(number of frames sample), zero padding will
+                         be done for generating last frame.
 
     Returns:
-            array: stacked_frames-Array of frames of size
-                (number_of_frames x frame_len).
+            array: stacked_frames-Array of frames of size (number_of_frames x frame_len).
 
     """
 
-    # Check dimension
+    ## Check dimension
     assert sig.ndim == 1, "Signal dimention should be of the format of (N,) but it is %s instead" % str(sig.shape)
 
     # Initial necessary values
     length_signal = sig.shape[0]
-    frame_sample_length = int(
-        np.round(
-            sampling_frequency *
[…]
