+++ speechpy/processing.py
@@ -90,7 +90,8 @@ def stack_frames(
     """
 
     # Check dimension
-    assert sig.ndim == 1, "Signal dimention should be of the format of (N,) but it is %s instead" % str(sig.shape)
+    assert sig.ndim == 1, "Signal dimention should be of the format of (N,) but it is %s instead" % str(
+        sig.shape)
 
     # Initial necessary values
     length_signal = sig.shape[0]
@@ -103,8 +104,9 @@ def stack_frames(
     # Zero padding is done for allocating space for the last frame.
     if zero_padding:
         # Calculation of number of frames
-        numframes = 1 + \
-            int(math.ceil((length_signal - frame_sample_length) / frame_stride))
+        numframes = (1
+                     + int(math.ceil((length_signal
+                                      - frame_sample_length) / frame_stride)))
 
         # Zero padding
         len_sig = int((numframes - 1) * frame_stride + frame_sample_length)
@@ -114,8 +116,9 @@ def stack_frames(
     else:
         # No zero padding! The last frame which does not have enough
         # samples(remaining samples <= frame_sample_length), will be dropped!
-        numframes = 1 + \
-            int(math.floor((length_signal - frame_sample_length) / frame_stride))
+        numframes = 1
+        + int(math.floor((length_signal
+                          - frame_sample_length) / frame_stride))
 
         # new length
         len_sig = int((numframes - 1) * frame_stride + frame_sample_length)
@@ -142,16 +145,21 @@ def stack_frames(
 
 
 def fft_spectrum(frames, fft_points=512):
-    """This function computes the one-dimensional n-point discrete Fourier Transform (DFT) of a real-valued
-    array by means of an efficient algorithm called the Fast Fourier Transform (FFT). Please refer to
-    https://docs.scipy.org/doc/numpy/reference/generated/numpy.fft.rfft.html for further details.
+    """This function computes the one-dimensional n-point discrete Fourier
+    Transform (DFT) of a real-valued array by means of an efficient algorithm
+    called the Fast Fourier Transform (FFT). Please refer to
+    https://docs.scipy.org/doc/numpy/reference/generated/numpy.fft.rfft.html
+    for further details.
 
     Args:
         frames (array): The frame array in which each row is a frame.
-        fft_points (int): The length of FFT. If fft_length is greater than frame_len, the frames will be zero-padded.
+        fft_points (int): The length of FFT. If fft_length is greater
+        than frame_len, the frames will be zero-padded.
 
     Returns:
-            array: The fft spectrum - If frames is an num_frames x sample_per_frame matrix, output will be num_frames x FFT_LENGTH.
+            array: The fft spectrum.
+            If frames is an num_frames x sample_per_frame matrix, output
+            will be num_frames x FFT_LENGTH.
     """
     SPECTRUM_VECTOR = np.fft.rfft(frames, n=fft_points, axis=-1, norm=None)
     return np.absolute(SPECTRUM_VECTOR)
@@ -162,10 +170,13 @@ def power_spectrum(frames, fft_points=512):
 
     Args:
         frames (array): The frame array in which each row is a frame.
-        fft_points (int): The length of FFT. If fft_length is greater than frame_len, the frames will be zero-padded.
+        fft_points (int): The length of FFT. If fft_length is greater than
+        frame_len, the frames will be zero-padded.
 
     Returns:
-            array: The power spectrum - If frames is an num_frames x sample_per_frame matrix, output will be num_frames x fft_length.
+            array: The power spectrum:
+            If frames is an num_frames x sample_per_frame matrix, output
+            will be num_frames x fft_length.
     """
     return 1.0 / fft_points * np.square(fft_spectrum(frames, fft_points))
[…]
