+++ speechpy/feature.py
@@ -1,22 +1,24 @@
 """feature module.
 
-This module demonstrates documentation for the main speech features that the 
+This module demonstrates documentation for the main speech features that the
 package is aimed to extract as well as the required elements.
 
 
 Attributes:
 
     filterbanks: Compute the Mel-filterbanks
-                 The filterbanks must be created for extracting speech features such as MFCC.
-    
+                 The filterbanks must be created for extracting
+                 speech features such as MFCC.
+
     mfcc: Extracting Mel Frequncy Cepstral Coefficient feature.
 
     mfe: Extracting Mel Energy feature.
 
     lmfe: Extracting Log Mel Energy feature.
 
-    extract_derivative_feature: Extract the first and second derivative features. This finction, 
-                                directly use the ``derivative_extraction`` function in the ``processing`` module.
+    extract_derivative_feature: Extract the first and second derivative
+        features. This finction, directly use the ``derivative_extraction``
+        function in the ``processing`` module.
 
 """
 
@@ -27,39 +29,56 @@ from scipy.fftpack import dct
 from . import functions
 
 
-def filterbanks(num_filter, coefficients, sampling_freq, low_freq=None, high_freq=None):
-    """Compute the Mel-filterbanks. Each filter will be stored in one rows. The columns correspond to fft bins.
+def filterbanks(
+        num_filter,
+        coefficients,
+        sampling_freq,
+        low_freq=None,
+        high_freq=None):
+    """Compute the Mel-filterbanks. Each filter will be stored in one rows.
+    The columns correspond to fft bins.
 
     Args:
         num_filter (int): the number of filters in the filterbank, default 20.
         coefficients (int): (fftpoints//2 + 1). Default is 257.
-        sampling_freq (float): the samplerate of the signal we are working with. Affects mel spacing.
+        sampling_freq (float): the samplerate of the signal we are working
+            with. It affects mel spacing.
         low_freq (float): lowest band edge of mel filters, default 0 Hz
-        high_freq (float): highest band edge of mel filters, default samplerate/2
+        high_freq (float): highest band edge of mel filters,
+            default samplerate/2
 
     Returns:
-           array: A numpy array of size num_filter x (fftpoints//2 + 1) which are filterbank
+           array: A numpy array of size num_filter x (fftpoints//2 + 1)
+               which are filterbank
     """
     high_freq = high_freq or sampling_freq / 2
     low_freq = low_freq or 300
-    assert high_freq <= sampling_freq / 2, "High frequency cannot be greater than half of the sampling frequency!"
+    s = "High frequency cannot be greater than half of the sampling frequency!"
+    assert high_freq <= sampling_freq / 2, s
     assert low_freq >= 0, "low frequency cannot be less than zero!"
 
-    ######################################################
-    ########### Computing the Mel filterbank #############
-    ######################################################
-
+    # Computing the Mel filterbank
     # converting the upper and lower frequencies to Mels.
-    # num_filter + 2 is because for num_filter filterbanks we need num_filter+2 point.
-    mels = np.linspace(functions.frequency_to_mel(low_freq), functions.frequency_to_mel(high_freq), num_filter + 2)
+    # num_filter + 2 is because for num_filter filterbanks we need
+    # num_filter+2 point.
+    mels = np.linspace(
+        functions.frequency_to_mel(low_freq),
+        functions.frequency_to_mel(high_freq),
+        num_filter + 2)
 
-    # we should convert Mels back to Hertz because the start and end-points should be at the desired frequencies.
+    # we should convert Mels back to Hertz because the start and end-points
+    # should be at the desired frequencies.
     hertz = functions.mel_to_frequency(mels)
 
     # The frequency resolution required to put filters at the
     # exact points calculated above should be extracted.
[…]
