### 2018-06-29 #749 @astorfi
@ThomasA The modules and functions related to SpeechPy, have been modified to be ai aligned with [PEP 8 -- Style Guide for Python Code](https://www.python.org/dev/peps/pep-0008/#fn-hi).
