+++ speechpy/feature (copy).py
@@ -1,206 +1,0 @@
-"""feature module.
-
-This module demonstrates documentation for the main speech features that the 
-package is aimed to extract as well as the required elements.
-
-
-Attributes:
-
-    filterbanks: Compute the Mel-filterbanks
-                 The filterbanks must be created for extracting speech features such as MFCC.
-    
-    mfcc: Extracting Mel Frequncy Cepstral Coefficient feature.
-
-    mfe: Extracting Mel Energy feature.
-
-    lmfe: Extracting Log Mel Energy feature.
-
-    extract_derivative_feature: Extract the first and second derivative features. This finction, 
-                                directly use the ``derivative_extraction`` function in the ``processing`` module.
-
-"""
-
-from __future__ import division
-import numpy as np
-from . import processing
-from scipy.fftpack import dct
-from . import functions
-
-
-def filterbanks(num_filter, coefficients, sampling_freq, low_freq=None, high_freq=None):
-    """Compute the Mel-filterbanks. Each filter will be stored in one rows. The columns correspond to fft bins.
-
-    Args:
-        num_filter (int): the number of filters in the filterbank, default 20.
-        coefficients (int): (fftpoints//2 + 1). Default is 257.
-        sampling_freq (float): the samplerate of the signal we are working with. Affects mel spacing.
-        low_freq (float): lowest band edge of mel filters, default 0 Hz
-        high_freq (float): highest band edge of mel filters, default samplerate/2
-
-    Returns:
-           array: A numpy array of size num_filter x (fftpoints//2 + 1) which are filterbank
-    """
-    high_freq = high_freq or sampling_freq / 2
-    low_freq = low_freq or 300
-    assert high_freq <= sampling_freq / 2, "High frequency cannot be greater than half of the sampling frequency!"
-    assert low_freq >= 0, "low frequency cannot be less than zero!"
-
-    ######################################################
-    ########### Computing the Mel filterbank #############
-    ######################################################
-
-    # converting the upper and lower frequencies to Mels.
-    # num_filter + 2 is because for num_filter filterbanks we need num_filter+2 point.
-    mels = np.linspace(functions.frequency_to_mel(low_freq), functions.frequency_to_mel(high_freq), num_filter + 2)
-
-    # we should convert Mels back to Hertz because the start and end-points should be at the desired frequencies.
-    hertz = functions.mel_to_frequency(mels)
-
-    # The frequency resolution required to put filters at the
-    # exact points calculated above should be extracted.
-    #  So we should round those frequencies to the closest FFT bin.
-    freq_index = (np.floor((coefficients + 1) * hertz / sampling_freq)).astype(int)
-
-    # Initial definition
-    filterbank = np.zeros([num_filter, coefficients])
-
-    # The triangular function for each filter
-    for i in range(0, num_filter):
-        left = int(freq_index[i])
-        middle = int(freq_index[i + 1])
-        right = int(freq_index[i + 2])
-        z = np.linspace(left, right, num=right - left + 1)
-        filterbank[i, left:right + 1] = functions.triangle(z, left=left, middle=middle, right=right)
-
-    return filterbank
-
-def mfcc(signal, sampling_frequency, frame_length=0.020, frame_stride=0.01,num_cepstral =13,
-             num_filters=40, fft_length=512, low_frequency=0, high_frequency=None, dc_elimination=True):
-    """Compute MFCC features from an audio signal.
-
-    Args:
-
-         signal (array): the audio signal from which to compute features. Should be an N x 1 array
-         sampling_frequency (int): the sampling frequency of the signal we are working with.
-         frame_length (float): the length of each frame in seconds. Default is 0.020s
-         frame_stride (float): the step between successive frames in seconds. Default is 0.02s (means no overlap)
[…]
