+++ speechpy/processing (copy).py
@@ -1,301 +1,0 @@
-# -*- coding: utf-8 -*-
-"""Processing module for signal processing operations.
-
-This module demonstrates documentation for the signal processing function which are 
-required as internal computations in the package.
-
-
-Attributes:
-
-    preemphasis: Preemphasising on the signal. This is a preprocessing step.
-    
-    stack_frames: Create stacking frames from the raw signal.
-
-    fft_spectrum: Calculation of the Fast Fourier Transform.
-
-    power_spectrum: Power Spectrum calculation.
-
-    log_power_spectrum: Log Power Spectrum calculation.
-
-    derivative_extraction: Calculation of the derivative of the extracted featurs.
-
-    cmvn: Cepstral mean variance normalization. This is a post processing operation.
-
-    cmvnw: Cepstral mean variance normalization over the sliding window. This is a post processing operation.
-
-"""
-
-__license__ = "MIT"
-__author__ = " $Amirsina Torfi"
-__docformat__ = 'reStructuredText'
-
-import decimal
-import numpy as np
-import math
-
-
-# 1.4 becomes 1 and 1.6 becomes 2. special case: 1.5 becomes 2.
-def round_half_up(number):
-    return int(decimal.Decimal(number).quantize(decimal.Decimal('1'), rounding=decimal.ROUND_HALF_UP))
-
-
-def preemphasis(signal, shift=1, cof=0.98):
-    """preemphasising on the signal.
-
-    Args:
-        signal (array): The input signal.
-        shift (int): The shift step.
-        cof (float): The preemphasising coefficient. 0 equals to no filtering.
-
-    Returns:
-           the pre-emphasized signal.
-    """
-
-    rolled_signal = np.roll(signal, shift)
-    return signal - cof * rolled_signal
-
-def stack_frames(sig, sampling_frequency, frame_length=0.020, frame_stride=0.020, filter=lambda x: np.ones((x,)),
-                 zero_padding=True):
-    """Frame a signal into overlapping frames.
-
-    Args:
-        sig (array): The audio signal to frame of size (N,).
-        sampling_frequency (int): The sampling frequency of the signal.
-        frame_length (float): The length of the frame in second.
-        frame_stride (float): The stride between frames.
-        filter (array): The time-domain filter for applying to each frame. By default it is one so nothing will be changed.
-        zero_padding (bool): If the samples is not a multiple of frame_length(number of frames sample), zero padding will
-                         be done for generating last frame.
-
-    Returns:
-            array: stacked_frames-Array of frames of size (number_of_frames x frame_len).
-
-    """
-
-    ## Check dimension
-    assert sig.ndim == 1, "Signal dimention should be of the format of (N,) but it is %s instead" % str(sig.shape)
-
-    # Initial necessary values
-    length_signal = sig.shape[0]
-    frame_sample_length = int(np.round(sampling_frequency * frame_length))  # Defined by the number of samples
-    frame_stride = float(np.round(sampling_frequency * frame_stride))
-
-    # Zero padding is done for allocating space for the last frame.
-    if zero_padding:
-        # Calculation of number of frames
-        numframes = 1 + int(math.ceil((length_signal - frame_sample_length) / frame_stride))
-
-        # Zero padding
-        len_sig = int((numframes - 1) * frame_stride + frame_sample_length)
-        additive_zeros = np.zeros((len_sig - length_signal,))
-        signal = np.concatenate((sig, additive_zeros))
-
-    else:
-        # No zero padding! The last frame which does not have enough
-        # samples(remaining samples <= frame_sample_length), will be dropped!
[…]
