### 2018-06-30 6f7aced
add references
