+++ speechpy/tests/test_speechpy.py
@@ -14,6 +14,15 @@ mu, sigma = 0, 0.1 # mean and standard deviation
 signal = np.random.normal(mu, sigma, 1000000)
 fs = 16000
 
+ # Generating stached frames with SpeechPy
+frame_length = 0.02
+frame_stride = 0.02
+frames = processing.stack_frames(signal, sampling_frequency=fs,
+                                  frame_length=frame_length,
+                                  frame_stride=frame_stride,
+                                  filter=lambda x: np.ones((x,)),
+                                  zero_padding=True)
+
 class Test_Methods_Exists(object):
     def test_processing(self):
         
@@ -57,22 +66,31 @@ class Test_Processing(object):
        assert signal_preemphasized.shape == signal.shape
        
     def test_stack_frames(self):
-        
-        # Generating stached frames with SpeechPy
-        frame_length = 0.02
-        frame_stride = 0.02
-        frames = processing.stack_frames(signal, sampling_frequency=fs,
-                                          frame_length=frame_length,
-                                          frame_stride=frame_stride,
-                                          filter=lambda x: np.ones((x,)),
-                                          zero_padding=True)
-        
+                
+        # Direct calculation using numpy
         window = int(np.round(frame_length * fs))
         step = int(np.round(frame_stride * fs))
         all_frames = (int(np.ceil((signal.shape[0]
                                       - window) / step)))
         
+        # Shape matching of stacked frames
         assert all_frames == frames.shape[0]
+    
+    def test_cmvn(self):
+        
+        normalized_feature = processing.cmvn(frames, variance_normalization=True)
+        
+        # Shape match
+        assert normalized_feature.shape == frames.shape
+        
+        # Check the std and mean of the output vector
+        assert np.allclose(np.mean(normalized_feature,axis=0), np.zeros((1,normalized_feature.shape[1])))
+        assert np.allclose(np.std(normalized_feature,axis=0), np.ones((1,normalized_feature.shape[1])))
+        
+        
+
+        
+        
         
         
         
