+++ requirements.txt
@@ -2,3 +2,4 @@ numpy
 scipy
 speechpy
 python-coveralls
+pytest
