### 2018-06-30 e7901fb
add references
