### 2018-06-30 #749 @astorfi
@ThomasA  Tests has been added as well. Although there is room for improvement as well. However, I added the tests for different parts for continuous integration as well. Example usage has been separated.
