### 2018-06-30 3711a8e
Run black formatter on codebase
