+++ speechpy/__init__.py
@@ -1,2 +1,3 @@
+# flake8: noqa
 from . import feature
 from . import processing
