+++ .travis.yml
@@ -1,17 +1,19 @@
 language: python
 python:
-
   - "2.7"
   - "3.4"
   - "3.5"
+  - "3.6"
 
 # command to install dependencies
 install:
   - pip install -r requirements.txt
   - pip install coveralls
   - pip install codecov
+  - pip install flake8
 
 script:
+  - flake8 . --count --max-line-length=127 --show-source --statistics
   - coverage run --omit=*.virtualenvs*,*virtualenv* example/test_package.py test
   - coverage run --omit=*.virtualenvs*,*virtualenv* example/test_local.py test
   - pytest tests/
@@ -22,4 +24,3 @@ after_success:
   - codecov
 
 sudo: enabled
-dist: trusty

[…]
