+++ speechpy/feature.py
@@ -156,7 +156,8 @@ def mfcc(
 def mfe(signal, sampling_frequency, frame_length=0.020, frame_stride=0.01,
         num_filters=40, fft_length=512, low_frequency=0, high_frequency=None):
     """Compute Mel-filterbank energy features from an audio signal.
-
+    
+    Args:
          signal (array): the audio signal from which to compute features.
              Should be an N x 1 array
          sampling_frequency (int): the sampling frequency of the signal
