### 2018-07-01 e2494c6
add references
