### 2018-07-01 19f6fdc
add references
