### 2018-07-03 #749 @ThomasA
> Example usage: Glancing through the documentation, I did not find any usage examples, but the test scripts can be considered not-too-well-documented examples. A few usage examples that were easier to find and documented with a bit of explanatory text would be nice to have.

The former test scripts are now used as usage examples and likewise shown in the README. I think this is fine now.
