### 2018-07-03 #749 @ThomasA
> Automated tests: There are tests available in the 'tests' folder. These are in my opinion more like usage examples and not proper tests in the sense that they do not check anything. They run some example commands and display the output. In my opinion, proper automated tests should check the output and determine whether it is correct, resulting in clearly 'passed' or 'failed' test cases.

There are now automated tests using `pytest` in the 'tests' folder. They are not overly thorough (the coverage seems a bit low and the tests mostly check details like correct shape of outputs), but some testing is better than none and now it has a testing framework set up that more tests can easily be added to.
