### 2018-07-03 #749 @ThomasA
As far as I am concerned, I am done with the review and can **recommend publication**.
If you like, @astorfi , I can nitpick and style-comb the repository a bit more, but I am trying not to get too intrusive :-)
