### 2018-07-03 #749 @astorfi
@ThomasA  Thank you so much for your suggestions, modifications, and review. You are always welcome to propose changes in the repo.
