### 2018-07-06 9754a14
add references
