### 2018-07-12 #749 @kczimm
I'm sorry I'm just seeing this. I don't know the required timeline but I can finish a review of this package over the weekend.
