### 2018-07-12 #749 @arfon
@kczimm - great, thanks! Please go ahead and carry out the review in this thread. The timeline you propose would be perfect.
