### 2018-07-16 #749 @kczimm
I see the automated tests in the tests directory, however the coverage is minimal, as was previously mentioned. Although I wouldn't say more is required for recommendation, I strongly urge increasing the test coverage.
