### 2018-07-16 #749 @astorfi
@kczimm  Thank you for your suggestions.

About the first one:
[test_package.py](https://github.com/astorfi/speechpy/blob/master/example/test_package.py) is designed to exactly perform the job of [test_local.py](https://github.com/astorfi/speechpy/blob/master/example/test_local.py). [test_package.py](https://github.com/astorfi/speechpy/blob/master/example/test_package.py) file also requires **a pip install**. Both are available in the [example](https://github.com/astorfi/speechpy/tree/master/example) folder.
