### 2018-07-16 #749 @astorfi
@kczimm  About the tests, you are absolutely correct. However, as the package is currently a feature extractor, it is not trivial to add tests for accuracy tests. I tried to add tests as I make sure the package can be tested for minimal run requirements. It also tests the outputs to be in the correct format.
