### 2018-07-18 #749 @kczimm
@astorfi Thanks for the explanations. Based on these and the explanations you've already provided to @ThomasA, I've completed my review and can recommend publication. Good work.
