+++ paper/paper.bib
@@ -1,3 +1,90 @@
+@inproceedings{hirsch2000aurora,
+  title={The Aurora experimental framework for the performance evaluation of speech recognition systems under noisy conditions},
+  author={Hirsch, Hans-G{\"u}nter and Pearce, David},
+  booktitle={ASR2000-Automatic Speech Recognition: Challenges for the new Millenium ISCA Tutorial and Research Workshop (ITRW)},
+  year={2000}
+}
+
+@book{guyon2008feature,
+  title={Feature extraction: foundations and applications},
+  author={Guyon, Isabelle and Gunn, Steve and Nikravesh, Masoud and Zadeh, Lofti A},
+  volume={207},
+  year={2008},
+  publisher={Springer}
+}
+
+@article{furui1986speaker,
+  title={Speaker-independent isolated word recognition using dynamic features of speech spectrum},
+  author={Furui, Sadaoki},
+  journal={IEEE Transactions on Acoustics, Speech, and Signal Processing},
+  volume={34},
+  number={1},
+  pages={52--59},
+  year={1986},
+  publisher={IEEE}
+}
+
+@book{yu2016automatic,
+  title={AUTOMATIC SPEECH RECOGNITION.},
+  author={Yu, Dong and Deng, Li},
+  year={2016},
+  publisher={Springer}
+}
+
+@book{rabiner1993fundamentals,
+  title={Fundamentals of speech recognition},
+  author={Rabiner, Lawrence R and Juang, Biing-Hwang},
+  volume={14},
+  year={1993},
+  publisher={PTR Prentice Hall Englewood Cliffs}
+}
+
+@article{campbell1997speaker,
+  title={Speaker recognition: A tutorial},
+  author={Campbell, Joseph P},
+  journal={Proceedings of the IEEE},
+  volume={85},
+  number={9},
+  pages={1437--1462},
+  year={1997},
+  publisher={IEEE}
+}
+
+
+@inproceedings{deng2013recent,
+  title={Recent advances in deep learning for speech research at Microsoft},
+  author={Deng, Li and Li, Jinyu and Huang, Jui-Ting and Yao, Kaisheng and Yu, Dong and Seide, Frank and Seltzer, Michael and Zweig, Geoff and He, Xiaodong and Williams, Jason and others},
+  booktitle={Acoustics, Speech and Signal Processing (ICASSP), 2013 IEEE International Conference on},
+  pages={8604--8608},
+  year={2013},
+  organization={IEEE}
+}
+
+@inproceedings{lee2009unsupervised,
+  title={Unsupervised feature learning for audio classification using convolutional deep belief networks},
+  author={Lee, Honglak and Pham, Peter and Largman, Yan and Ng, Andrew Y},
+  booktitle={Advances in neural information processing systems},
+  pages={1096--1104},
+  year={2009}
+}
+
+@inproceedings{yu2011improved,
+  title={Improved bottleneck features using pretrained deep neural networks},
+  author={Yu, Dong and Seltzer, Michael L},
+  booktitle={Twelfth Annual Conference of the International Speech Communication Association},
+  year={2011}
+}
+
+@article{giannakopoulos2015pyaudioanalysis,
+  title={pyAudioAnalysis: An Open-Source Python Library for Audio Signal Analysis},
+  author={Giannakopoulos, Theodoros},
+  journal={PloS one},
+  volume={10},
+  number={12},
+  year={2015},
+  publisher={Public Library of Science}
+}
+
 @article{torfi2017text,
   title={Text-independent speaker verification using 3d convolutional neural networks},
[…]
