+++ README.rst
@@ -14,8 +14,8 @@
     :target: https://codecov.io/gh/astorfi/speechpy
 .. image:: https://badge.fury.io/py/speechpy.svg
     :target: https://badge.fury.io/py/speechpy
-.. image:: https://zenodo.org/badge/DOI/10.5281/zenodo.1066373.svg
-   :target: https://doi.org/10.5281/zenodo.1066373
+.. image:: https://zenodo.org/badge/DOI/10.5281/zenodo.810391.svg
+    :target: https://doi.org/110.5281/zenodo.810391
 
 .. _SpeechPy Official Project Documentation: http://speechpy.readthedocs.io
 
