### 2018-07-23 #749 @astorfi
@arfon  I updated the version in Zenodo as well as the PyPi.
Please check to see if it is suitable.
Thanks
