### 2018-07-23 #749 @arfon
Looks like the Zenodo DOI (https://doi.org/110.5281/zenodo.810391) isn't resolving yet... Perhaps it takes a few hours to register?
