### 2018-07-23 fd24dc3
add references
