### 2018-07-23 #749 @astorfi
@arfon I realized that I made a mistake and I put the DOI as [https://doi.org/110.5281/zenodo.810391](https://doi.org/110.5281/zenodo.810391) instead of [https://doi.org/10.5281/zenodo.810391](https://doi.org/10.5281/zenodo.810391). I think it is solved now.
