### 2018-07-23 #749 @arfon
@astorfi - It looks like the citations aren't being compiled into the paper as you're not citing them directly. You can read how to do that [here](http://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html).
