### 2018-07-23 #749 @astorfi
@arfon I see the citations in the article proof. I am not quite sure if I got your point.
