### 2018-07-24 #749 @arfon
> @arfon I see the citations in the article proof. I am not quite sure if I got your point.

Sorry for any confusion. I was working with an earlier clone of your repo.

BTW, I did need to make a small change to your `paper.md` to get the images working: https://github.com/astorfi/speechpy/pull/25
