### 2018-07-24 #749 @whedon
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
